Resolve API-797 "Add filters"

Closes API-797

See merge request ci-api/admin-api!75

+++ internal/config/config.go
@@ -69,6 +69,6 @@ type Solvimon struct {
 }
 
 type Pagination struct {
-	MaxPageSize     uint16
-	DefaultPageSize uint16
+	MaxPageSize     uint
+	DefaultPageSize uint
 }
